Added highlightcircle animation, refactored some methods into base-graph, and made graphtraversal more smooth
